Changed the fundamental parts of the bot. Added prefix to the config.json file. Added SQLite to perform database actions (it saves the tables in a .SQLITE file).
Added light security to prevent other bots from using our prefix and causing a spamloop. Also added that the bot will not listen to chat commands if the defined prefix is not used.

Added the ability to add a role to a certain user. User needs to have a certain role (see line 25/26).